refactor(coediting): add docstrings and consolidate all coedit code into lib/coediting/
